Add concept of serverUnique columns to avoid duplicate db entries.
This change supports adding uniqueness constraints to fields other than
the primary key field.  This is helpful when your primary key is a value
that only exists in your local database but other columns contain keys
that exist on the server.  In cases were you are reading data from the
server and attempt to insert an entry that violates a server key constraint,
you want to be able reload your instance with the existing row entry that
matches the given server key.  You may now define with column(s) define your
server key by annotating columns as having a constraint `.serverUnique`.