Refactor insert/update handling
If you have a constraint violation on in an Insert, the library will
now detect the failed insert and instead perform an Update on the db
row that was a duplicate for all of the non-key columns.  It will
then read that row from the db and throw an exception that contains
the newly read object from the db.